jobs/pyproject.toml
[project]
name = "jobs"
version = "0.1.0"
description = "Cloud Run Jobs — LightGBM training"
requires-python = ">=3.12"
dependencies = [
    "common",
    "google-auth>=2.36",
    "httpx>=0.27",
    "lightgbm>=4.5",
    "numpy>=2.0",
    "pandas>=2.2",
    "wandb>=0.17",
]

[project.scripts]
train = "training.entrypoints.cli:main"
rank-train = "training.entrypoints.rank_cli:main"
embed = "training.entrypoints.embed_cli:main"
meili-sync = "meili_sync.entrypoints.sync_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/training"]
